BackendView now supports groups. Opening, print, running, testing, and viewing readmes all look for the handins of any member of the group, and then open the one that is found.
Also fixes the enabled/disabled state of the print code and print rubric buttons. They are now enabled when something of what is selected can be printed, even if not everything selected can be printed. On print, only what can actually be printed is sent off to printing - to prevent errors for nonexistant files.

Continues progress on issue 88.