Resolves Issue 198 : Allocator's allocations placed behind interfaces : Modifies Allocator so that what it allocates can be modified for the purposes of a JUnit test. Because each allocation could rely on any number of other allocations, and this could occur in the constructor, allocations are placed within a SingletonAllocation, which is just a fancy type of thunk. (Oh, how a Racket-style macro would make for a lot less code.)
Classes that are provided by the Allocator are now all behind interfaces. All public methods in the interfaces have the javadoc comments, not the implementations. (This means the IDE should provide them in either case, and duplication would just be asking for them to fall out of sync.)

Classes that make use of the Allocator have not been modified. These changes have not affected how a non-allocated class interacts with the Allocator.

An example that makes use of EasyMock and the ability to swap out what the Allocator allocates, see UserUtilitiesTest.

NOTE: RubricMananger should be RubricManager. It has been misnamed for a very long time. Trying to rename the class in this revision caused foodcritic to reject the diff, saying it could not find RubricManager. After this commit is approved, I will follow it up with a commit that just renames RubricMananger to RubricManager. : Verified that cakehat still executes properly. Created and ran a JUnit test to verify that the Allocator can be modified. : http://cakehat.cs.brown.edu/foodcritic/r/137/