resolves issue none : updates method to clear DB to also recreate the DB if none exists : now tables are first dropped then recreated. before the data was just being removed. this is better since it will reduce our file size and also allow new in memory DBs to be created with the same method.
the /trunk/db folder was also removed since it's data was outdated and the db template file is now nolonger needed.

methods for the distlookup view were also added due to the fact that they were in the same file as the changes I needed to commit. I can remove them if needed but it will be a pain. : tried with cs000. was running locally so it could not add tas or students back. the DB structure was inspected manually with sqlite3 command line tools.