Resolves Issue 222 : Throw errors from native calls : Native calls now throw exceptions. Exceptions are thrown if either there is an error in native code or if the information cannot be retrieved (for instance if the group or login do not exist). Code that calls these functions now either handle these exceptions or throw them higher up the call chain. The general approach taken was that exceptions are thrown by utility or service classes, and handled by their users (typically UI based classes such as the frontend or a backend view). That way the error message that is given to ErrorView can be specific to the use of the service or utility function.
Note: While previously looking up the members of a group that did not exist (for instance cs000ta on a non-department machine) would return an empty list, it now throws an exception because the group. Similarly, looking up the real name for a login that does not exist will thrown an exception instead of returning null. That is, native calls that silently failed before, now very explicitly fail. : All native methods were confirmed to throw exceptions when they failed and to return the correct information otherwise.  : http://cakehat.cs.brown.edu/foodcritic/r/148/