Resolution of Issues 208 & 217 : Create directories and files with correct permissions and group owner : Adds the ability to natively change group (chgrp is the *nix utility used to do this, although the corresponding native function is actually chown). This is integrated with correctly changing permissions into the FileSystemServices interface and FileSystemServicesImpl class which allow for creating directories and "sanitizing" created files such that they should have proper permissions and group ownership (the TA group). Because these methods throw exceptions, exception handling is added throughout.
In the process a few other changes were made. ArchiveUtilities now uses File instead of String (specifying the file path), this was more convenient because of the use of File elsewhere. FileSystemUtilities's chmodFile and chmodDirectory have been combined into chmodDefault which detects if it is operating on a file or directory. (Although neither method should typically be used, instead the sanitize method of FileSystemServices is preferable because it also sets the group ownership.)

Note: These changes make it necessary for the cs000ta group (or if you are testing with another course, that course's ta group) to actually exist on your system. If you do not do so then you will be bombarded with ErrorViews because the grading directory workspace, and subdirectories for each assignment, are created from both the frontend and backend in response to list selections. Doing so on OS X is very straightforward. (This will be added to the wiki later, but for now just follow here.) Open System Preferences, go to Accounts. Click the "+" in the lower left, you may need to unlock to do this. For the dialog that appears, under "New" choose group and then type in "cs000ta." The list of the left will now include "cs000ta," add your account to this group. : Confirmed that exceptions are thrown when the group does not exist. Confirmed no exceptions are thrown when the group does exist. Confirmed that the group and permissions for new directories and files are properly set, including recursively. Confirmed that makeDirectories(...) properly creates directories recursively and sets permissions recursively. : http://cakehat.cs.brown.edu/foodcritic/r/150/