resolves issue 165, 220 : DB refactor to remove TA table and remove the need to make unique ids for TAs, Students, and Assignments in the DB : Many if not all of the DB calls can now be done without joins which will speed up DB access and make the code much easier to read, update, and understand. The TA table has been removed entirely. The assignment and assignment part tables have also been removed. This make it so that the DB is now only responsible for grade related information storage. There is no duplication in other areas.
The get ID methods of both Assignment and Part of temporary and can be changed if need.

I will try to convert the existing DBs to work with this new code. If that is not possible we may need to rebuild the DBs which will be a pain. Old DB files will NOT work at all with this code. : ran the one :-( unit test for the DB and tried a few tasks with the frontend and backend. Extensive testing is hard since everything has been effected and unit tests are light at best.