Issue 230 : Improved file copying support : Overhauls copying support for files and directories. There are numerous helper functions involved in copying. Despite that, some of the methods are quite long due to the extensive exception handling and generation. Copying allows for the options of overwrite existing files (but not existing directories - instead their contents will be merged) and for preserving last modified date. If copying fails for any reason then all files copied so far will be deleted so that it is not left in an inconsistent state.
FileSystemUtilities does all of the heavy lifting of actually performing the copying. FileSystemServices has similar methods for copying that also after copying change the group on each file created from a result of copying to have the TA group be the group owner. This follows the pattern that utility classes be distinct from cakehat. File modifications are not changed, this is because unlike files cakehat creates (database, rubrics, etc.) the files copied may need executable permission (perhaps for a tester). : Extensive tests of copying files around my filesystem were performed. Copying a single files (up to ~80MB in size) was performed for a variety of files. Copying various directories was done, including merging into an existing directory structure. Copies were done with and without overwrite permission to ensure that overwriting occurred when specified and raised an exception when not. Preserving last modified date was also shown.

Important note: while the last modified date will be preserved, the creation date will be of the time the copy was performed. This means the last modified date will predate the creation date for the file. The creation date is what is shown in Finder for OS X 10.5. The only code in cakehat that relies on last modified date is determining the early/ontime/late status of handins, and currently we do not copy those files with cakehat. However, if we were, this copy support would be compatible with it. : http://cakehat.cs.brown.edu/foodcritic/r/168/