Issue 151 : PDF viewer and improved text editor support: Adds support for PDF opening with acroread, kpdf, evince, and xpdf. By default acroread is used because it has tabs, and all of the other programs open a separate window for each file. Which program is used is customizable via a property.
Overhauls text editor support so that many other editors besides kate may be used. The default is kate, but supported editors are: kate, gedit, bluefish, nedit, emacs, kwrite, vi, vim, nano, and pico. Support for other editors can easily be added if desired. The editor to be used for the action can be specified. Instead of, or additionally, the user's EDITOR environment variable can be used to determine which editor to use. If that editor is not supported or not set then either the specified or default editor will be used.

Removes the methods in ExternalProcessesUtilities that opened kate and kdpf as they are far too specific.

Note: Both of these actions are in a class called EditorActions. The class will be renamed ApplicationActions before commit, but foodcritic was unhappy about the diff produced from the renaming. (The Terminal action was added to this class as it fits in nicely.) : Tested that generation of the commands is correct and that the commands are a valid way to open files for that editor/viewer. : http://cakehat.cs.brown.edu/foodcritic/r/177/