Issue 231 : Adds a Java test mode that copies the contents of a specified directory into the root of the unarchived handin directory. It then compiles the code, and runs a specified main class - presumably (but does not have to be) a main class that was copied over.
Fixes a variety of small bugs and improves the word on some other actions for Java and MATLAB. : Tested the gfx tester against a copy of Othello. Ensured that copying occurs properly and does not proceed when files exist with the same names. : http://cakehat.cs.brown.edu/foodcritic/r/179/