Issue None : Lab checkoffs now done with Java : Lab checkoffs are now done with pure Java. The only non-Java part is the cakehat_labCheckOff script which calls into the cakehat jar.
Grades are entered directly into the database. Grades are entered for the entire group. Validation is performed, and issue are reported to the terminal.

CourseInfo was split into CourseInfo and ConfigurationInfo so that all path methods are distinct from parsing the configuration file. This is needed so that the database's location can be determined without parsing the configuration file. For now the methods forward throw to reduce merge conflicts with Jonathan's code. (From updated diff: Remove old lab importing / updating code.

Changes the structure of the checkoff code a bit to be more abstracted such that tests can be written against it without needing to interact with the command line.

Added a method to GeneralUtilities, not currently used, that silences System.err - this could be used to silence the error message the DB prints when initialized.) : Confirmed that checking off works. Confirmed that attempted overwriting prompts the user. Confirmed that user must exist, must be enabled, lab number must exist, and that grade given does not exceed maximum points for the lab. Several JUnit tests written : http://cakehat.cs.brown.edu/foodcritic/r/200/