Issue None : Fixes issues with spaces in file paths : Fixes issues in opening files, Java class paths, and printing with lpr due to spaces in file paths. These issues arise from external calls.
There are issues with external actions and they have yet to be resolved.
MATLAB may have issues with spaces in paths - this has not yet been tested. : Confirmed opening files and Java class paths works properly with spaces in the path. Did not test lpr command, but as this in an issue with interacting with the external console and not the commands themselves there is no reason lpr should operate any differently than a call to nano or kate. : http://cakehat.cs.brown.edu/foodcritic/r/205/