Resolves Issue None : Removes hardcoded paths to icons. : Removes hard-coded paths to images such that they are now loaded throw the IconLoader class. This removes over 60 hard-coded paths. This will allow for refactoring the gradesystem package to the cakehat package without breaking every single image path.
Note: There are still three remaining "hard-coded" paths in EmailView because these paths are generated by the GUI Builder and I know of no way to tell it to load an Icon in the manner using an arbitrary method. : Confirmed icons appear properly in the frontend and backend. : http://cakehat.cs.brown.edu/foodcritic/r/228/