Resolves Issue None : Shows submitted status in frontend : Shows with a green check mark the groups and distributable parts which have had their grades submitted. If a group has been submitted then a green check mark precedes the group name. If for a distributable part all groups have had their grades submitted, then a green check mark precedes the displayed name. These visual indications update in response to submitting grades.
The motivation for this change is to make it much clearer what remains to be graded, particularly when grading is being completed in batches as can occur in cs4, cs15 final project grading, and I am sure other situations.

Considerable update. Now all assigned groups are loaded asynchronously using a background thread and then placed onto the UI event queue as necessary. This makes the frontend extremely responsive while able to provide this additional information.

Groups that are submitted have a check mark and use plain gray text. Groups that have had work done to their rubrics are displayed in bold gray text. Groups with no work done are displayed in bold black text. These groups are sorted (top to bottom): submitted, work done on rubric, no work done. Within these categories they are sorted alphabetically by group name.

Distributable Parts that have all of their groups submitted have a check mark and are in plain gray text. If this criteria is not met then they are displayed in black bold text. Distributable Parts sorting has not been altered.

These statuses automatically update when saving a rubric or submitting grades. All information can be updated by performing a refresh (File -> Refresh or Ctrl + R). This is a complete refresh of all data so that if a group is assigned for a part that was previously not assigned then it will appear. Previously to make this appear cakehat would have had to be restarted.

The submit button's actions (submit, print grds, and email) were switched to using a background thread to improve UI responsiveness. In the process a bug was found and fixed relating to GRD generation. (GRDs were being generated if grades were submitted as opposed to if GRD files were to be attached to emails.) : Confirmed that the visualization appears correctly while changing between assignments and groups. Confirmed that the visual update occurs correctly after submitting grades. No bugs encountered during this manual testing. : http://cakehat.cs.brown.edu/foodcritic/r/232/