Resolves Issue None : RubricGMLWriter only chmods file it has created : Currently the RubricGMLWriter will chmod a file it is overwriting. If it does not own the file it is overwriting then the call to chmod will fail. It is unclear as to how this code previously functioned without issue.
RubricGMLWriter has been modified to only chmod a file if the file is being created as a result of the write. : Jonathan will do this on the department machines to verify proper functionality. I cannot adequately test it on my own machine with just 1 user account. : http://cakehat.cs.brown.edu/foodcritic/r/241/