Resolves Issue None : Handles uncaught exceptions and changes to ErrorView : Currently uncaught runtime exceptions, such as NullPointerException, have their stack traces to the terminal. If the user is running cakehat in a graphical mode (this should be all situations except for being 'su'-ed while checking off a lab) then an ErrorView is brought up. If in a non-graphical mode then the stack trace is printed to the terminal.
When in developer mode (cakehat started with the window that allows selecting between the backend and frontend) or having started cakehat without going through main class (as during testing or a unit test), ErrorView will print the stack trace to the terminal in addition to make itself appear. This is to aid debugging so that the stack trace can be clicked on in Netbeans to easily navigate to the problematic code. : Confirmed by creating NullPointerExceptions with intentionally faulty code that has now been removed. : http://cakehat.cs.brown.edu/foodcritic/r/244/