Resolves Issue None : Blacklist overhaul : The current blacklist UI was usable for the both the backend and frontend. It was designed for the backend, and looked less than optimal for the frontend. The new UI operates only in the frontend, and there is no longer any backend blacklist UI.
FrontendView now has two additional corresponding methods: disableFrame() and enableFrame()
disableFrame() disables the menu and content pane and then creates an overlay in the glasspane that always sits ontop of the content pane, but is normally transparent. Content is placed inside of a panel which visually appears inside of a rounded rectangle that overlaps with the top of the screen (see screen shot). The code is written in a generalized way so that other UIs could make use of it - such as grade submission. : Confirmed that black lists are updated properly and persist between restarts of cakehat.