Resolves Issue None : Java handin - supply run arguments : Handed in Java programs (but not demos) can be run with grader supplied run arguments.
In the process, removed the ability to specify which main should be run. The functionality already was that if there is only one main, to just run it without asking. Even if a main was specified, it was still being verified that it exists. There really is no need for the ability to specify a main, and it just complicated the code (and would have become more complicated when adding this feature). : Confirmed this functionality works by trying all of the possible options. Tested against a Clock handin.