Resolves Issue None : Drag and Drop JList : Extends GenericJList to create DnDList, a generic drag and drop JList. In the process generalizes GenericJList a bit to allow for subclasses of it to specify what model they are using. Generalizes the generic data models being used to have both immutable and mutable models. (GenericJList and GenericJComboBox still used immutable models.)
Refactors some of the method names in GenericJList to be more consistently named and improves some comments.

The DnDList is not yet used anywhere in cakehat, but I plan to make use of it configuration modification interface. It has been written in a very generalized manner, so there are likely other places in cakehat that could benefit from it. : Considerable testing has been done with the test code present at the end of the DnDList class inside of its mainline. (This code can be removed prior to a commit.)

Additionally, GenericJLists such as those in the Frontend, Backend, Manual Distributor, and Blacklist have been tested and found unaffected by the changes. : http://cakehat.cs.brown.edu/foodcritic/r/281/