Resolves Issue 183, 184, 185 : Ports charts off GUI builder : Converts AssignmentChartPanel, StatHistView, and StudentChartPanel. The StatHistView UI was simplified to have a drop down list of "By Assignment", "By Assignment Part", and "By Student" as those were the only options prior - but they were split across a drop down list and two mutually exclusive buttons. The conversion off the GUI builder was very basic - essentially only GUI code was modified.
Note: The charts can take a very long time to load, a new issue (#269), has been opened for this. An initial educated guess is this is because of the enormous number of database calls being made. : Confirmed that charts appear as they did before. : http://cakehat.cs.brown.edu/foodcritic/r/295/