Resolve Issue 242 : Recalculates handin status when granting extension for group already distributed : Currently it is possible to grant an extension to a group for an assignment that has already been distributed and there will be no effect unless handin statuses were recalculated. Now that scenario causes the handin status to be recalculated.
In implementing this bug fix, several other things were changed. A bug was found and fixed with the days of the month being incorrectly determined for the initial month. The class had an inner class IntegerField that very closely matched support.ui.IntegerField, it was changed to using that. In the process, IntegerField was modified to be more accommodating to this use case. All uses of JComboBox, which resulted in casting, were switched over GenericJComboBox.

No UI changes made. : Confirmed that groups not distributed functioned as before and that for groups already distributed the appropriate GradingService's method was called. : http://cakehat.cs.brown.edu/foodcritic/r/305/