Resolves Issue None : Removes unused important and unnecessary white space the beginning and end of files.
Replaces the built-in test dependency on JUnit to being an explicit dependency on JUnit with a junit-4.8.2.jar in testlib. This is better because it ensures that everyone developing cakehat on any version of Netbeans (and in the future perhaps Eclipse or IntelliJ) uses the exact same version. Removes dependencies previously used by GUI Builder code but no longer needed :
Confirmed that all JUnit tests still run. (Some of them failed both before and after these changes, this needs to be resolved - separately.) Confirmed that cakehat still compiles. : http://cakehat.cs.brown.edu/foodcritic/r/307/