Resolves issue 238 : New EmailAccount with encapsulated Java KeyStore and adds documentation : Adds a new cakehat.email package that currently just has two files - EmailAccount.java and brown_cs_email.jks. The jks file is what we have previously been calling a "cert" file, although it turns out that what we were using was a Java KeyStore file that incorporated a Brown CS certification file (crt) into it. The details of all of this are documented in EmailAccount.
This package will get additional classes added to it, but I wanted to get this through code review and checked in first as it's certainly going to be the most complicated the part. Those classes will include a manager of some sort that can gracefully handle the situation when the email info has not yet been configured and hold the notification email addresses. : A lot of manual testing by sending emails to myself. Tested both running from Netbeans and from the cakehat.jar. : http://cakehat.cs.brown.edu/foodcritic/r/422/