Resolves issue none : EmailAccount without using a Java KeyStore : Configures the socket factory used by mail to accept all hosts it connects to which means there is now no need to use a Java KeyStore. This means we are no longer validating that we are actually talking to the Brown CS email server - we are vulnerable to man in the middle attacks. I think that's fine - email clients (like Outlook) by default don't validate the server they are talking to.
If we do want to continue using the Java KeyStore it looks like I can directly configure the MailSSLSocketFactory so that there will not be an need to extract the jks file from the jar. However this solution is permanent while the keystore solution needs a new keystore in 2015 when the certificate it includes expires. : Tested sending mail to myself. : http://cakehat.cs.brown.edu/foodcritic/r/423/