Resolves issue 180 : GradeReportView : This is a grade report view in all its glory. Currently the dataservices calls are pretty inefficient and for enough data it can take a really long time to open - after this is checked in I think it is definitely worth it to create more efficient ultrabulk methods to retrieve the needed data (the CSV exporter can also make use of them).
This code contains a *very* small amount of multithreading. Emails are sent on a separate thread because it can take ~1-2 seconds per student to generate and send the email (the vast majority of that time is sending the email, not generating it). The key aspect of this is that all of the loaded data is put into immutable data structures because they are guaranteed to be visible to the email sending thread. : Extensive manual testing by sending to students whose email address was set as my personal email address. : http://cakehat.cs.brown.edu/foodcritic/r/513/