Resolves issue 400 : Top level window changes : Changes to having only one top level window at a time:  - DeveloperModeView (initial window shown during development which allows for selecting one of the top three)  - AdminView  - GraderView  - ConfigManagerView
Some modal dialog windows are created without an owner so they are also top-level windows during their time on screen. This should be changed in the future, but it's tolerable for now because the application cannot be terminated normally while a modal dialog is on screen. The ErrorView is not modal and is also a top-level window, this should be improved as part of the overall error management improvement (issue 297).

Aside from the above exceptions, all secondary windows are now owned to the appropriate top level window. As a result when the top level window is disposed, windows owned to it will also be closed. There are no longer any explicit or implicit (via setDefaultCloseOperation(EXIT_ON_CLOSE) close operations. As such, listeners for window closing operations will be able to run as expected. After all code has finished running cakehat will terminate as expected.

This solves issue 400 where grading sheets windows launched by the grader view were not saving when the grader view was closed. This likely also addresses other unknown bugs of a similar nature. : Manually confirmed that issue 400 was resolved by these changes. : http://cakehat.cs.brown.edu/foodcritic/r/524/