Resolves issue none : Manage Groups button : Adds a button to the admin view for managing groups. It will only appear when a group assignment is selected. Also cleans up some of the padding code used to space out buttons.
When the button is pressed it shows a dialog saying the feature is not yet implemented. Wil will need to hook his manage groups UI into the code at that point. : Tested manually to confirm button appears when it should and that padding is correct. : http://cakehat.cs.brown.edu/foodcritic/r/532/