Resolves issue none : This does not directly address any bugs, but the generalizations made by this is a prerequisite for supporting digital handbacks and for actions to ignore files created by test actions.
Makes part actions and their descriptions more unified. Part actions now takes in an ActionContext and return an ActionResult - neither of which do very much right now, but in the future will be used for actions to communicate any files they created as well as other information.
As part of the over all generalization:
 - Viewing a readme is now a non-user configurable part action, instead of being an integrated capability of Part.
 - Grading guide is now a user configurable part action just like demo, instead of being a column in the part table. This allows a course to swap out the implementation and for us to easily extend this capability in the future.
 - Various code simplification in Part and its related classes : Manual testing : http://foodcritic.cakehat.org/r/20/