Resolves issue 406, resolves issue 407 : Centralized cakehat : Changes cakehat to run from a centralized location and take in the course as an argument. Creates a cakehat script that lives at /contrib/bin/cakehat. This script finds the latest version of the cakehat jar which will live at /contrib/projects/cakehat/releases/latest/cakehat.jar.
For a course to start using cakehat all it takes is for a HTA to type in: cakehat -c course -m config. This causes cakehat to run for that course and then proceeds to launch the configuration manager. In this case the user is brought initially to a new tab added to the config manager called Scripts. This tab allows for creating course specific convenience scripts for running cakehat and cakehat in enter grade mode. For instance, cs000 could create a cs000_cakehat script. Then for a TA to run cakehat for cs000 they would just type cs000_cakehat and they'd be good to go. To run in admin mode a user would type in cs000_cakehat -m admin, or for config cs000_cakehat -m config. There's no requirement for courses to create these scripts, they can always use the cakehat script directly. : Manually tested and passes all automation tests. : http://foodcritic.cakehat.org/r/36/