Resolves issue 417 : Notifies courses of file permissions issues : FileSystemUtilities getFiles(...) methods now throws FilePermissionException if it cannot read the contents of a directory.
Adds interface ThrowableReporter in the cakehat.logging package. ErrorReporter has a list of ThrowableReporters and the first one which says it can handle the thrown exception gets to report it. Right now there are just two implementations of ThrowableReporter, the first one handles FilePermissionException if all of the files associated with the exception is not in the course's .cakehat directory. The second reporter is adapted from the current implementation of error reporting and shows the error view (or command line experience if there is no UI available). : Manually tested. : http://foodcritic.cakehat.org/r/75/